+++ app/application.jsx
@@ -2,24 +2,22 @@ import 'stylesheets/application';
 import React from 'react';
 import { render } from 'react-dom';
 import { Router, Route, Redirect, browserHistory } from 'react-router';
-import AppLayout from 'components/application/layout';
-import Main from 'components/main/view';
-import AboutLayout from 'components/about/layout';
-import About from 'components/about/view';
-import Detailed from 'components/about/detailed';
+import Application from 'components/application';
+import Main from 'components/main';
+import About from 'components/about';
+import Article from 'components/article';
 import { requireAuth } from 'helpers/routes';
 
 render((
   <Router history={ browserHistory }>
-    <Route component={ AppLayout }>
+    <Route component={ Application }>
       <Route path="/" component={ Main }/>
       <Route
         path="about"
-        component={ AboutLayout }
+        component={ About }
         onEnter={ requireAuth }
       >
-        <Route path="extended" component={ About }/>
-        <Route path="extended/:id" component={ Detailed }/>
+        <Route path="extended/:id" component={ Article }/>
       </Route>
       <Redirect from="*" to="/"/>
     </Route>