+++ app/components/article/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Router from 'react-router';
 
-export default class Detailed extends React.Component {
+export default class Article extends React.Component {
   static propTypes = {
     params: React.PropTypes.shape({
       id: React.PropTypes.string