+++ app/lib/request-auth.js
@@ -2,7 +2,7 @@ import deepAssign from 'deep-assign';
 import { request } from 'lib/request';
 import session from 'services/session';
 
-export function requestAuth(url, params) {
+export function requestAuth(url, params, queryParams) {
   const defaultParams = {
     headers: {
       'X-User-Token': session.token,
@@ -10,5 +10,5 @@ export function requestAuth(url, params) {
     }
   };
 
-  return request(url, deepAssign({}, defaultParams, params));
+  return request(url, deepAssign({}, defaultParams, params), queryParams);
 }