+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Rewrite all scripts without usage of gulp. Remove gulp from project.
+([#27](https://github.com/fs/react-base/pull/27))
 - Upgrade Node.js to 6.2.x
 ([#26](https://github.com/fs/react-base/pull/26))
 - Lock core-decorators package version to avoid tests errors.